Fix findAll python interface to be usable for Python users
Fixing a few issues of the python interface to findAll:
* The output parameter could not be created as there was not a named
SWIG wrap of type std::list<NavDataPtr>
* The interface required an output parameter which is atypical for
Python. The find interface is already adjusted with a SWIG typemap to
move the output parameter to be a regular return parameter.
* The output NavData objects could not be cast to their dynamic type
by python users making the underlying data unreachable.

The python version of the findAll method previously returned a list of NavData pointers which
would be unusable to Python users. Once in Python-land, the objects
cannot be cast to their dynamic type and preventing any useful logic on
the objects.

The find function's python interface already behaves to convert the
return NavData object into their dynamic type before passing the object to the
Python user. I adapted this solution to the findAll interface such that
the entire list of NavData objects are converted to their dynamic type
for the python user.

This fix makes findAll usable for python users.